feat: Add loading UI to auth
