+++ client/src/api/auth.js
@@ -11,12 +11,12 @@ let options = {
 
 export async function registerUser(userData) {
   const url = `${BASE_URL}/@me/register`;
-
   const res = await fetch(url, {
     ...options,
     method: "POST",
     body: JSON.stringify(userData),
   });
+
   const data = await res.json();
 
   if (!res.ok) {
@@ -54,10 +54,13 @@ export async function updateUser(userData) {
       Authorization: `Bearer ${token}`,
     },
   });
+
   const data = await res.json();
+
   if (!res.ok) {
     throw new Error(data.message || "Something went wrong");
   }
+
   return data;
 }
 
@@ -79,6 +82,7 @@ export async function createQuiz(userData) {
   if (!res.ok) {
     throw new Error(data.message || "Something went wrong");
   }
+
   return data;
 }
 
@@ -99,6 +103,7 @@ export async function getQuizById(id) {
   if (!res.ok) {
     throw new Error(data.message || "Something went wrong");
   }
+
   return data;
 }
 
@@ -118,6 +123,7 @@ export async function deleteQuizById(id) {
   if (!res.ok) {
     throw new Error(data.message || "Something went wrong");
   }
+
   return data;
 }
 
@@ -138,5 +144,6 @@ export async function getAllQuizzes(page = 1) {
   if (!res.ok) {
     throw new Error(data.message || "Something went wrong");
   }
+
   return data;
 }
